data type change

+++ CRT_FOP_DSA/src/Recursion/Q4.java
@@ -8,13 +8,13 @@ public class Q4 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.print("Enter a number : ");
-        int num = sc.nextInt();
-        int res = sum(num);
+        long num = sc.nextInt();
+        long res = sum(num);
         System.out.println(s);
 
     }
-static int s= 0;
-    public static int sum(int n){
+static long s= 0;
+    public static int sum(long n){
         if(n==0) return 0;
         s = sum += n%10;
         return sum(n/10);
